+++ vortex-array/src/array/composite/array.rs
@@ -7,9 +7,10 @@ use vortex_error::VortexResult;
 use vortex_schema::{CompositeID, DType};
 
 use crate::array::composite::{find_extension, CompositeExtensionRef, TypedCompositeArray};
-use crate::array::{Array, ArrayRef, Encoding, EncodingId, EncodingRef, ENCODINGS};
+use crate::array::{Array, ArrayRef};
 use crate::compress::EncodingCompression;
 use crate::compute::ArrayCompute;
+use crate::encoding::{Encoding, EncodingId, EncodingRef, ENCODINGS};
 use crate::formatter::{ArrayDisplay, ArrayFormatter};
 use crate::impl_array;
 use crate::serde::{ArraySerde, BytesSerde, EncodingSerde};
