+++ vortex-array/src/array/mod.rs
@@ -1,10 +1,7 @@
 use arrow_buffer::Buffer;
 use std::any::Any;
 use std::fmt::{Debug, Display, Formatter};
-use std::hash::{Hash, Hasher};
 use std::sync::Arc;
-
-use linkme::distributed_slice;
 
 use vortex_error::{VortexError, VortexResult};
 use vortex_schema::{DType, Nullability};
@@ -19,7 +16,6 @@ use crate::array::sparse::{SparseArray, SparseEncoding};
 use crate::array::struct_::{StructArray, StructEncoding};
 use crate::array::varbin::{VarBinArray, VarBinEncoding};
 use crate::array::varbinview::{VarBinViewArray, VarBinViewEncoding};
-use crate::compress::EncodingCompression;
 use crate::compute::as_arrow::AsArrowArray;
 use crate::compute::as_contiguous::AsContiguousFn;
 use crate::compute::cast::CastFn;
@@ -31,7 +27,7 @@ use crate::compute::search_sorted::SearchSortedFn;
 use crate::compute::take::TakeFn;
 use crate::compute::ArrayCompute;
 use crate::formatter::{ArrayDisplay, ArrayFormatter};
-use crate::serde::{ArraySerde, EncodingSerde};
+use crate::serde::ArraySerde;
 use crate::stats::Stats;
 use crate::validity::{ArrayValidity, Validity};
 
@@ -122,6 +118,7 @@ macro_rules! impl_array {
     };
 }
 
+use crate::encoding::EncodingRef;
 pub use impl_array;
 
 impl ArrayCompute for ArrayRef {
@@ -380,69 +377,6 @@ pub fn check_validity_buffer(validity: Option<&ArrayRef>, expected_len: usize) -
     }
 
     Ok(())
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct EncodingId(&'static str);
-
-impl EncodingId {
-    pub const fn new(id: &'static str) -> Self {
-        Self(id)
-    }
-
-    #[inline]
-    pub fn name(&self) -> &'static str {
-        self.0
-    }
-}
-
-impl Display for EncodingId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        Display::fmt(self.0, f)
-    }
-}
-
-pub trait Encoding: Debug + Send + Sync + 'static {
-    fn id(&self) -> EncodingId;
-
-    /// Whether this encoding provides a compressor.
-    fn compression(&self) -> Option<&dyn EncodingCompression> {
-        None
-    }
-
-    /// Array serialization
-    fn serde(&self) -> Option<&dyn EncodingSerde> {
-        None
-    }
-}
-
-impl Display for dyn Encoding {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.id())
-    }
-}
-
-pub type EncodingRef = &'static dyn Encoding;
-
-impl PartialEq<Self> for EncodingRef {
-    fn eq(&self, other: &Self) -> bool {
-        self.id() == other.id()
-    }
-}
-
-impl Eq for EncodingRef {}
-
-impl Hash for EncodingRef {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id().hash(state)
-    }
-}
-
-#[distributed_slice]
-pub static ENCODINGS: [EncodingRef] = [..];
-
-pub fn find_encoding(id: &str) -> Option<EncodingRef> {
-    ENCODINGS.iter().find(|&x| x.id().name() == id).cloned()
 }
 
 #[derive(Debug, Clone)]
