+++ vortex-array/src/compress.rs
@@ -12,9 +12,10 @@ use crate::array::constant::ConstantArray;
 use crate::array::sparse::SparseEncoding;
 use crate::array::struct_::{StructArray, StructEncoding};
 use crate::array::varbin::VarBinEncoding;
-use crate::array::{Array, ArrayKind, ArrayRef, Encoding, EncodingRef, ENCODINGS};
+use crate::array::{Array, ArrayKind, ArrayRef};
 use crate::compute;
 use crate::compute::scalar_at::scalar_at;
+use crate::encoding::{Encoding, EncodingRef, ENCODINGS};
 use crate::formatter::display_tree;
 use crate::sampling::stratified_slices;
 use crate::stats::Stat;
